+++ app/src/main/java/me/tatarka/bindingcollectionadapter/sample/Listeners.java
@@ -3,18 +3,8 @@ package me.tatarka.bindingcollectionadapter.sample;
 /**
  * Created by evan on 6/3/15.
  */
-public class Listeners {
-    private MyViewModel viewModel;
+public interface Listeners {
+    void onAddItem();
 
-    public Listeners(MyViewModel viewModel) {
-        this.viewModel = viewModel;
-    }
-   
-    public void onAddItem() {
-        viewModel.addItem();
-    }
-
-    public void onRemoveItem() {
-        viewModel.removeItem();
-    }
+    void onRemoveItem();
 }